kafka-docker/kafka-data-saver/requirements.txt
six
python-dotenv
kafka-python